fix(/student/mine/lib;/types/questions): add a tip of mine/lib page to push students to share
