+++ src/pages/Student/Mine/Lib/index.module.less
@@ -3,6 +3,13 @@
 
 .lib-container {
   position: relative;
+  .share-tips {
+    position: absolute;
+    right: 0;
+    top: -26px;
+    color: @font-green;
+    font-weight: bold;
+  }
 
   .back_college {
     position: fixed;
